refactor: extract inline tests from runtime, types, gym_history
- Extract runtime/mod.rs inline tests → runtime/tests_mod.rs
- Extract engineer_loop/types.rs inline tests → engineer_loop/tests_types_inline.rs
- Convert gym_history.rs to gym_history/mod.rs directory module
- Extract gym_history inline tests → gym_history/tests.rs
- Add #[cfg(test)] mod declarations in respective mod.rs files
